Add data cleanup job based on retention settings

+++ SocrataCache.Jobs/DownloadPendingDatasetsJob.cs
@@ -1,8 +1,7 @@
 using System.IO.Compression;
 using Microsoft.Extensions.Logging;
-using SocrataCache.Config;
-
 using Quartz;
+using SocrataCache.Config;
 using SocrataCache.Managers.Models;
 
 namespace SocrataCache.Jobs;
@@ -13,7 +12,8 @@ public class DownloadPendingDatasetsJob : IJob
     private readonly Managers.DatasetManager _datasetManager;
     private readonly ILogger<FreshDatasetLookupJob> _logger;
 
-    public DownloadPendingDatasetsJob(Config.Config config, Managers.DatasetManager datasetManager, ILogger<FreshDatasetLookupJob> logger)
+    public DownloadPendingDatasetsJob(Config.Config config, Managers.DatasetManager datasetManager,
+        ILogger<FreshDatasetLookupJob> logger)
     {
         _config = config;
         _datasetManager = datasetManager;
@@ -23,7 +23,7 @@ public class DownloadPendingDatasetsJob : IJob
     public async Task Execute(IJobExecutionContext context)
     {
         _logger.LogInformation("Running Download Pending Datasets Job");
-        
+
         foreach (var resource in _config.GetResources())
         {
             var pendingDataset = await _datasetManager.GetDatasetByStatus(DatasetStatus.Pending, resource.ResourceId);
@@ -39,36 +39,31 @@ public class DownloadPendingDatasetsJob : IJob
                 _logger.LogInformation($"Downloading dataset for {resource.ResourceId}");
 
                 await _datasetManager.UpdateDatasetStatus(pendingDataset.DatasetId, DatasetStatus.Downloading);
-                
+
                 using var httpClient = new HttpClient();
-                
+
                 var datasetsDirectory = Env.DownloadsRootPath.Value;
-                
-                if (!Directory.Exists(datasetsDirectory))
-                {
-                    Directory.CreateDirectory(datasetsDirectory);
-                }
-                
+
+                if (!Directory.Exists(datasetsDirectory)) Directory.CreateDirectory(datasetsDirectory);
+
                 var fileNameDownload = $"{resource.ResourceId}-{pendingDataset.DatasetId}.csv";
                 var fileNameExisting = $"{resource.ResourceId}.csv";
-                var fileNameObsolete = $"{resource.ResourceId}-obsolete.csv";
-                
+
                 var fileNameDownloadCompressed = $"{resource.ResourceId}-{pendingDataset.DatasetId}.csv.gz";
                 var fileNameExistingCompressed = $"{resource.ResourceId}.csv.gz";
-                var fileNameObsoleteCompressed = $"{resource.ResourceId}-obsolete.csv.gz";
 
                 var filePathDownload = $"{datasetsDirectory}/{fileNameDownload}";
                 var filePathExisting = $"{datasetsDirectory}/{fileNameExisting}";
-                var filePathObsolete = $"{datasetsDirectory}/{fileNameObsolete}";
                 var filePathDownloadCompressed = $"{datasetsDirectory}/{fileNameDownloadCompressed}";
                 var filePathExistingCompressed = $"{datasetsDirectory}/{fileNameExistingCompressed}";
-                var filePathObsoleteCompressed = $"{datasetsDirectory}/{fileNameObsoleteCompressed}";
-                
+
                 await _datasetManager.UpdateDatasetStatus(pendingDataset.DatasetId, DatasetStatus.Downloading);
 
                 var baseUri = _config.GetBaseUri();
-                
-                await using (var stream = await httpClient.GetStreamAsync(resource.GetDownloadUrl(baseUri, await resource.GetColumns(baseUri))))
+
+                await using (var stream =
+                             await httpClient.GetStreamAsync(resource.GetDownloadUrl(baseUri,
+                                 await resource.GetColumns(baseUri))))
                 await using (var fileStream = File.Create(filePathDownload))
                 {
                     await stream.CopyToAsync(fileStream);
@@ -76,46 +71,28 @@ public class DownloadPendingDatasetsJob : IJob
 
                 _logger.LogInformation("Downloaded");
 
-                _logger.LogInformation("Deleting obsolete file");
-                if (File.Exists(filePathObsolete))
-                {
-                    File.Delete(filePathObsolete);
-                }
-
-                _logger.LogInformation("Make existing file obsolete");
-                if (File.Exists(filePathExisting))
-                {
-                    File.Delete(filePathExisting);
-                }
+                if (File.Exists(filePathExisting)) File.Delete(filePathExisting);
 
                 _logger.LogInformation("Completing download");
-                File.Move(filePathDownload, filePathExisting);
-                
+                File.Copy(filePathDownload, filePathExisting);
+
                 _logger.LogInformation("Compressing download");
-                
+
                 await using (var downloadFileStream = File.OpenRead(filePathExisting))
                 await using (var compressedFileStream = File.Create(filePathDownloadCompressed))
                 await using (var gzipStream = new GZipStream(compressedFileStream, CompressionLevel.Optimal))
                 {
                     await downloadFileStream.CopyToAsync(gzipStream);
                 }
-                
-                if (File.Exists(filePathObsoleteCompressed))
-                {
-                    File.Delete(filePathObsoleteCompressed);
-                }
-                
-                if (File.Exists(filePathExistingCompressed))
-                {
-                    File.Move(filePathExistingCompressed, filePathObsoleteCompressed);
-                }
-                
+
+                if (File.Exists(filePathExistingCompressed)) File.Delete(filePathExistingCompressed);
+
                 File.Move(filePathDownloadCompressed, filePathExistingCompressed);
-                
+
                 await _datasetManager.UpdateDatasetStatus(pendingDataset.DatasetId, DatasetStatus.Downloaded);
-                
+
                 _logger.LogInformation("Compressing complete");
-                
+
                 _logger.LogInformation("Checking if cleanup is needed for retention settings");
             }
             catch (Exception ex)
